componente deletado e ajuste na responsividade
